+++ docs/source/changes.md
@@ -39,6 +39,8 @@ authors:
 - `bids_root` and `deriv_root` are now converted to absolute paths to avoid
   running into issues caused by relative path specifications.
   ({{ gh(322) }} by {{ authors.hoechenberger }})
+- Solve issue 337: Warning when 0 EOG-related IC detected.
+  ({{ gh(?) }} by {{ authors.crsegerie }})
 
 ### Bug fixes